Refactor project setup to scope tvOS functions into extension
